refac: conteudo do puzzle movido para puzzle_page.dart
